+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Display from price only for variable products
 * Display price contained in attribute description (only if no voucher)
+* Rename "Your order" by "Your appointments" when appointment only
 
 ### 1.0.7: June 30th, 2020
 * Fix checking cart contents when on "order-pay" page